Moved ml models into modeling.py
